FCL-944 | split out search views
