+++ judgments/views/search/__init__.py
@@ -1,0 +1,4 @@
+from .advanced_search import AdvancedSearchView
+from .results import SearchResultsView
+
+__all__ = ["SearchResultsView", "AdvancedSearchView"]
